requirements.txt
colorama
web3
requests
python-dotenv
cryptography
eth-account
eth-utils